requirements.txt
flask
python-pptx
Pillow